Fixed: Stuttering touch/drag of timeline produced by mousewheel
2016: When playing around in Chrome of my Mac with the embedded timeline within our web system, I've noticed that the timeline begins heavily to stutter when previously the mouse wheel was used in any scrollable content.

The problem remains even exist when the URL is reloaded again. The only way to get a "smooth scrolling" of timeline back, is to close Chrome browser or to open a new tab.

You can reproduce it, when:

Open: http://visjs.org/examples/timeline/other/performance.html
draw 10000
Zoom out timeline to the max and touch and drag the timeline a lot to the left (it will scroll smooth all the time)
Now resize the browser window to a small size, so that a vertical scrollbar appears
Drag the scrollbar to the bottom without using the mouse wheel !!! (only by dragging the scrollbar)
Resize the browser window to big size and touch and drag timeline a lot to the left -> (new items will appear and the scrolling of timeline is smooth all the time - no problem at all)
Now Resize browser window to a very small window, so that a vertical scrollbar appears again
And now use only the mouse wheel to scroll the scrollbar to the bottom
Resize the browser window to a big size and touch and drag the timeline a lot to left (now the timeline will start heavily to stutter!!)
Reload URL -> zoom out to max-> scroll timeline to left -> it will start to stutter again after touch and drag the timeline
The problem was found when using Chrome on Unix (as ubuntu) and Mac.
That problem has nothing to do with resizing browser view, because stuttering appears in general, when using a mouse wheel for scrolling a scrollable element (e.g. combobox...) before.

https://github.com/almende/vis/issues/2023